requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

google_api_python_client == 2.12.0
google_auth_oauthlib == 0.4.4
praw == 7.3.0
protobuf == 3.16.0
requests == 2.25.1
